+++ tests/ArrayTest.php
@@ -39,6 +39,7 @@ class ArrayTest extends \PHPUnit_Framework_TestCase
         $array = objectToArray($obj);
         $this->assertInternalType('array', $array);
     }
+
     public function test_arrayToObject()
     {
         $array = [
@@ -52,8 +53,9 @@ class ArrayTest extends \PHPUnit_Framework_TestCase
 
         $array = [
             'foo' => 'bar',
-            'baz' => [ 'foo2' => 'bar2',
-                        'baz2' => 'ok',
+            'baz' => [
+                'foo2' => 'bar2',
+                'baz2' => 'ok',
             ],
         ];
         $obj = arrayToObject($array);
@@ -63,6 +65,7 @@ class ArrayTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals('bar2', $obj->baz->foo2);
         $this->assertEquals('ok', $obj->baz->baz2);
     }
+
     public function test_arrayToString()
     {
         $array = [
@@ -81,8 +84,8 @@ class ArrayTest extends \PHPUnit_Framework_TestCase
         $this->assertTrue(isNullOrEmptyArray([]));
         $this->assertFalse(isNullOrEmptyArray(['']));
         $this->assertFalse(isNullOrEmptyArray([' ']));
-        $this->assertFalse(isNullOrEmptyArray([1,2,3,4]));
-        $this->assertFalse(isNullOrEmptyArray(['1' => 1,'2' => 2,'3' => 3,'4' => 4]));
+        $this->assertFalse(isNullOrEmptyArray([1, 2, 3, 4]));
+        $this->assertFalse(isNullOrEmptyArray(['1' => 1, '2' => 2, '3' => 3, '4' => 4]));
     }
 
     public function test_isNotNullOrEmptyArray()
@@ -90,52 +93,112 @@ class ArrayTest extends \PHPUnit_Framework_TestCase
         $this->assertFalse(isNotNullOrEmptyArray([]));
         $this->assertTrue(isNotNullOrEmptyArray(['']));
         $this->assertTrue(isNotNullOrEmptyArray([' ']));
-        $this->assertTrue(isNotNullOrEmptyArray([1,2,3,4]));
-        $this->assertTrue(isNotNullOrEmptyArray(['1' => 1,'2' => 2,'3' => 3,'4' => 4]));
+        $this->assertTrue(isNotNullOrEmptyArray([1, 2, 3, 4]));
+        $this->assertTrue(isNotNullOrEmptyArray(['1' => 1, '2' => 2, '3' => 3, '4' => 4]));
     }
 
     public function test_isNullOrEmptyArrayKey()
     {
-        $this->assertTrue(isNullOrEmptyArrayKey([],'2',true));
-        $this->assertTrue(isNullOrEmptyArrayKey([''],'2',true));
-        $this->assertTrue(isNullOrEmptyArrayKey([' '],'2',true));
-        $this->assertFalse(isNullOrEmptyArrayKey([1,2,3,4],'2',true));
-        $this->assertFalse(isNullOrEmptyArrayKey(['1' => 1,'2' => 2,'3' => 3,'4' => 4],'2',true));
-        $this->assertTrue(isNullOrEmptyArrayKey(['1' => 1,'2' => '','3' => 3,'4' => 4],'2',false));
-        $this->assertTrue(isNullOrEmptyArrayKey(['1' => 1,'2' => '','3' => 3,'4' => 4],'2',true));
-        $this->assertFalse(isNullOrEmptyArrayKey(['1' => 1,'2' => ' ','3' => 3,'4' => 4],'2',false));
-        $this->assertTrue(isNullOrEmptyArrayKey(['1' => 1,'2' => ' ','3' => 3,'4' => 4],'2',true));
+        $this->assertTrue(isNullOrEmptyArrayKey([], '2', true));
+        $this->assertTrue(isNullOrEmptyArrayKey([''], '2', true));
+        $this->assertTrue(isNullOrEmptyArrayKey([' '], '2', true));
+        $this->assertFalse(isNullOrEmptyArrayKey([1, 2, 3, 4], '2', true));
+        $this->assertFalse(isNullOrEmptyArrayKey(['1' => 1, '2' => 2, '3' => 3, '4' => 4], '2', true));
+        $this->assertTrue(isNullOrEmptyArrayKey(['1' => 1, '2' => '', '3' => 3, '4' => 4], '2', false));
+        $this->assertTrue(isNullOrEmptyArrayKey(['1' => 1, '2' => '', '3' => 3, '4' => 4], '2', true));
+        $this->assertFalse(isNullOrEmptyArrayKey(['1' => 1, '2' => ' ', '3' => 3, '4' => 4], '2', false));
+        $this->assertTrue(isNullOrEmptyArrayKey(['1' => 1, '2' => ' ', '3' => 3, '4' => 4], '2', true));
     }
 
     public function test_isNotNullOrEmptyArrayKey()
     {
-        $this->assertFalse(isNotNullOrEmptyArrayKey([],'2',true));
-        $this->assertFalse(isNotNullOrEmptyArrayKey([''],'2',true));
-        $this->assertFalse(isNotNullOrEmptyArrayKey([' '],'2',true));
-        $this->assertTrue(isNotNullOrEmptyArrayKey([1,2,3,4],'2',true));
-        $this->assertTrue(isNotNullOrEmptyArrayKey(['1' => 1,'2' => 2,'3' => 3,'4' => 4],'2',true));
-        $this->assertFalse(isNotNullOrEmptyArrayKey(['1' => 1,'2' => '','3' => 3,'4' => 4],'2',false));
-        $this->assertFalse(isNotNullOrEmptyArrayKey(['1' => 1,'2' => '','3' => 3,'4' => 4],'2',true));
-        $this->assertTrue(isNotNullOrEmptyArrayKey(['1' => 1,'2' => ' ','3' => 3,'4' => 4],'2',false));
-        $this->assertFalse(isNotNullOrEmptyArrayKey(['1' => 1,'2' => ' ','3' => 3,'4' => 4],'2',true));
+        $this->assertFalse(isNotNullOrEmptyArrayKey([], '2', true));
+        $this->assertFalse(isNotNullOrEmptyArrayKey([''], '2', true));
+        $this->assertFalse(isNotNullOrEmptyArrayKey([' '], '2', true));
+        $this->assertTrue(isNotNullOrEmptyArrayKey([1, 2, 3, 4], '2', true));
+        $this->assertTrue(isNotNullOrEmptyArrayKey(['1' => 1, '2' => 2, '3' => 3, '4' => 4], '2', true));
+        $this->assertFalse(isNotNullOrEmptyArrayKey(['1' => 1, '2' => '', '3' => 3, '4' => 4], '2', false));
+        $this->assertFalse(isNotNullOrEmptyArrayKey(['1' => 1, '2' => '', '3' => 3, '4' => 4], '2', true));
+        $this->assertTrue(isNotNullOrEmptyArrayKey(['1' => 1, '2' => ' ', '3' => 3, '4' => 4], '2', false));
+        $this->assertFalse(isNotNullOrEmptyArrayKey(['1' => 1, '2' => ' ', '3' => 3, '4' => 4], '2', true));
     }
 
     public function test_array_get_key_value_safe()
     {
-        $this->assertEquals('default', array_get_key_value_safe([],'2','default'));
-        $this->assertEquals('default', array_get_key_value_safe([''],'2','default'));
-        $this->assertEquals('default', array_get_key_value_safe([' '],'2','default'));
-        $this->assertEquals(3, array_get_key_value_safe([1,2,3,4],'2','default'));
-        $this->assertEquals(2, array_get_key_value_safe(['1' => 1,'2' => 2,'3' => 3,'4' => 4],'2',2));
+        $this->assertEquals('default', array_get_key_value_safe([], '2', 'default'));
+        $this->assertEquals('default', array_get_key_value_safe([''], '2', 'default'));
+        $this->assertEquals('default', array_get_key_value_safe([' '], '2', 'default'));
+        $this->assertEquals(3, array_get_key_value_safe([1, 2, 3, 4], '2', 'default'));
+        $this->assertEquals(2, array_get_key_value_safe(['1' => 1, '2' => 2, '3' => 3, '4' => 4], '2', 2));
     }
 
     public function test_insert_at_top()
     {
-        $test_arr = array('a','b','c','d');
-        $test_arr_result = array(99,98,97,96,95,94,93,92,91,90);
-        for($i=0;$i<100; $i++) {
+        $test_arr = array('a', 'b', 'c', 'd');
+        $test_arr_result = array(99, 98, 97, 96, 95, 94, 93, 92, 91, 90);
+        for ($i = 0; $i < 100; $i++) {
             insert_at_top($test_arr, $i, 10);
         }
         $this->assertEquals($test_arr, $test_arr_result);
+    }
+
+    public function test_array_remove_columns()
+    {
+        $test_arr = [
+            0 => [
+                0 => 'abc',
+                1 => 123,
+                2 => 'aaaaa',
+            ],
+            1 => [
+                0 => 'def',
+                1 => 456,
+                2 => 'ddddd',
+            ],
+            2 => [
+                0 => 'ghi',
+                1 => 789,
+                2 => 'hhhhhhh',
+            ],
+        ];
+
+        $test_arr_result =
+            array
+            (
+                0 => array
+                (
+                    0 => 'abc',
+                    1 => 123,
+                ),
+                1 => array
+                (
+                    0 => 'def',
+                    1 => 456,
+                ),
+                2 => array
+                (
+                    0 => 'ghi',
+                    1 => 789,
+                )
+            );
+        $test_arr_result_first = [
+            0 => [
+                0 => 123,
+                1 => 'aaaaa',
+            ],
+            1 => [
+                0 => 456,
+                1 => 'ddddd',
+            ],
+            2 => [
+                0 => 789,
+                1 => 'hhhhhhh',
+            ],
+        ];
+
+        $this->assertEquals($test_arr_result, array_remove_columns($test_arr, 3));
+        $this->assertEquals($test_arr_result, array_remove_last_columns($test_arr));
+        $this->assertEquals($test_arr_result_first, array_remove_columns($test_arr, 1));
+        $this->assertEquals($test_arr_result_first, array_remove_first_columns($test_arr));
     }
 }