+++ tests/StringDataProvider.php
@@ -55,4 +55,34 @@ trait StringDataProvider
             '10, 4' => [10, 4, true],
         ];
     }
+
+    /**
+     * @return array
+     */
+    public function slugfyProvider()
+    {
+        return [
+            'null, null' => [null, null, 'TypeError'],
+            '\'\', null' => ['', null, 'TypeError'],
+            '\'\', \'\'' => ['', '', ''],
+            '0, \'-\'' => [0, '-', '0'],
+            '\' foo  bar \', \'-\'' => [' foo  bar ', '-', 'foo-bar'],
+            '\'foo -.-"-...bar\', \'-\'' => ['foo -.-"-...bar', '-', 'foo-bar'],
+            '\'another..& foo -.-"-...bar\', \'-\'' => ['another..& foo -.-"-...bar', '-', 'another-foo-bar'],
+            '\' Foo d\'Bar \', \'-\'' => [' Foo d\'Bar ', '-', 'foo-dbar'],
+            '\'A string-with-dashes\', \'-\'' => ['A string-with-dashes', '-', 'a-string-with-dashes'],
+            '\'Using strings like fòô bàř\', \'-\'' => ['Using strings like fòô bàř', '-', 'using-strings-like-foo-bar'],
+            '\'numbers 1234\', \'-\'' => ['numbers 1234', '-', 'numbers-1234'],
+            '\'śšşсσșςစſს\', \'-\'' => ['śšşсσșςစſს', '-', 'ssssssssss'],
+            '\'°₀۰\', \'-\'' => ['°₀۰', '-', '000'],
+            '\'Foo bar baz\', \'_\'' => ['Foo bar baz', '_', 'foo_bar_baz'],
+            '\'A_string with_underscores\', \'_\'' => ['A_string with_underscores', '_', 'a_string_with_underscores'],
+            '\'--   An odd__   string-_\', \'-\'' => ['--   An odd__   string-_', '-', 'an-odd-string'],
+            '\'A string / strong\', \'-\'' => ['A string / strong', '-', 'a-string-strong'],
+            '\'A string \ strong\', \'-\'' => ['A string \ strong', '-', 'a-string-strong'],
+            '\'foo           Bar\' (spaces U+2000 to U+200A), \'-\'' => ['foo           Bar', '-', 'foo-bar'],
+            '\'foo Bar\' (no-break space (U+00A0)), \'-\'' => ['foo Bar', '-', 'foo-bar'],
+            '\'foo𐍉Bar\' (some uncommon, unsupported character (U+10349)), \'-\'' => ['foo𐍉Bar', '-', 'foobar'],
+        ];
+    }
 }